backend/requirements.txt
Django==6.0.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
python-dotenv==1.0.1
Pillow==11.1.0
gunicorn==23.0.0
google-generativeai==0.8.3
PyPDF2==3.0.1
sentence-transformers==3.0.1
pinecone-client==5.0.0